table refactor for readability
