+++ crates/nu-table/src/lib.rs
@@ -1,5 +1,9 @@
 mod table;
+mod table_theme;
+mod textstyle;
 mod wrap;
 
-pub use table::{draw_table, StyledString, Table, TextStyle, Theme};
+pub use table::{draw_table, Table};
+pub use table_theme::TableTheme;
+pub use textstyle::{StyledString, TextStyle};
 pub use wrap::Alignment;
