+++ crates/nu-table/src/table.rs
@@ -1,5 +1,7 @@
+use crate::table_theme::TableTheme;
 use crate::wrap::{column_width, split_sublines, wrap, Alignment, Subline, WrappedCell};
-use nu_ansi_term::{Color, Style};
+use crate::{StyledString, TextStyle};
+use nu_ansi_term::Style;
 use nu_protocol::{Config, FooterMode};
 use std::collections::HashMap;
 use std::fmt::Write;
@@ -14,595 +16,15 @@ enum SeparatorPosition {
 pub struct Table {
     pub headers: Vec<StyledString>,
     pub data: Vec<Vec<StyledString>>,
-    pub theme: Theme,
-}
-
-#[derive(Debug, Clone)]
-pub struct StyledString {
-    pub contents: String,
-    pub style: TextStyle,
-}
-
-impl StyledString {
-    pub fn new(contents: String, style: TextStyle) -> StyledString {
-        StyledString { contents, style }
-    }
-
-    pub fn set_style(&mut self, style: TextStyle) {
-        self.style = style;
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct TextStyle {
-    pub alignment: Alignment,
-    pub color_style: Option<Style>,
-}
-
-impl TextStyle {
-    pub fn new() -> TextStyle {
-        TextStyle {
-            alignment: Alignment::Left,
-            color_style: Some(Style::default()),
-        }
-    }
-
-    pub fn bold(&self, bool_value: Option<bool>) -> TextStyle {
-        let bv = bool_value.unwrap_or(false);
-
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_bold: bv,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_bold(&self) -> bool {
-        self.color_style.unwrap_or_default().is_bold
-    }
-
-    pub fn dimmed(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_dimmed: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_dimmed(&self) -> bool {
-        self.color_style.unwrap_or_default().is_dimmed
-    }
-
-    pub fn italic(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_italic: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_italic(&self) -> bool {
-        self.color_style.unwrap_or_default().is_italic
-    }
-
-    pub fn underline(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_underline: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_underline(&self) -> bool {
-        self.color_style.unwrap_or_default().is_underline
-    }
-
-    pub fn blink(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_blink: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_blink(&self) -> bool {
-        self.color_style.unwrap_or_default().is_blink
-    }
-
-    pub fn reverse(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_reverse: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_reverse(&self) -> bool {
-        self.color_style.unwrap_or_default().is_reverse
-    }
-
-    pub fn hidden(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_hidden: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_hidden(&self) -> bool {
-        self.color_style.unwrap_or_default().is_hidden
-    }
-
-    pub fn strikethrough(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_strikethrough: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_strikethrough(&self) -> bool {
-        self.color_style.unwrap_or_default().is_strikethrough
-    }
-
-    pub fn fg(&self, foreground: Color) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                foreground: Some(foreground),
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn on(&self, background: Color) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                background: Some(background),
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn bg(&self, background: Color) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                background: Some(background),
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn alignment(&self, align: Alignment) -> TextStyle {
-        TextStyle {
-            alignment: align,
-            color_style: self.color_style,
-        }
-    }
-
-    pub fn style(&self, style: Style) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                foreground: style.foreground,
-                background: style.background,
-                is_bold: style.is_bold,
-                is_dimmed: style.is_dimmed,
-                is_italic: style.is_italic,
-                is_underline: style.is_underline,
-                is_blink: style.is_blink,
-                is_reverse: style.is_reverse,
-                is_hidden: style.is_hidden,
-                is_strikethrough: style.is_strikethrough,
-            }),
-        }
-    }
-
-    pub fn basic_center() -> TextStyle {
-        TextStyle::new()
-            .alignment(Alignment::Center)
-            .style(Style::default())
-    }
-
-    pub fn basic_right() -> TextStyle {
-        TextStyle::new()
-            .alignment(Alignment::Right)
-            .style(Style::default())
-    }
-
-    pub fn basic_left() -> TextStyle {
-        TextStyle::new()
-            .alignment(Alignment::Left)
-            .style(Style::default())
-    }
-
-    pub fn default_header() -> TextStyle {
-        TextStyle::new()
-            .alignment(Alignment::Center)
-            .fg(Color::Green)
-            .bold(Some(true))
-    }
-
-    pub fn default_field() -> TextStyle {
-        TextStyle::new().fg(Color::Green).bold(Some(true))
-    }
-
-    pub fn with_attributes(bo: bool, al: Alignment, co: Color) -> TextStyle {
-        TextStyle::new().alignment(al).fg(co).bold(Some(bo))
-    }
-
-    pub fn with_style(al: Alignment, style: Style) -> TextStyle {
-        TextStyle::new().alignment(al).style(Style {
-            foreground: style.foreground,
-            background: style.background,
-            is_bold: style.is_bold,
-            is_dimmed: style.is_dimmed,
-            is_italic: style.is_italic,
-            is_underline: style.is_underline,
-            is_blink: style.is_blink,
-            is_reverse: style.is_reverse,
-            is_hidden: style.is_hidden,
-            is_strikethrough: style.is_strikethrough,
-        })
-    }
-}
-
-impl Default for TextStyle {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Theme {
-    pub top_left: char,
-    pub middle_left: char,
-    pub bottom_left: char,
-    pub top_center: char,
-    pub center: char,
-    pub bottom_center: char,
-    pub top_right: char,
-    pub middle_right: char,
-    pub bottom_right: char,
-    pub top_horizontal: char,
-    pub middle_horizontal: char,
-    pub bottom_horizontal: char,
-    pub left_vertical: char,
-    pub center_vertical: char,
-    pub right_vertical: char,
-
-    pub separate_header: bool,
-    pub separate_rows: bool,
-
-    pub print_left_border: bool,
-    pub print_right_border: bool,
-    pub print_top_border: bool,
-    pub print_bottom_border: bool,
-}
-
-impl Theme {
-    #[allow(unused)]
-    pub fn basic() -> Theme {
-        Theme {
-            top_left: '+',
-            middle_left: '+',
-            bottom_left: '+',
-            top_center: '+',
-            center: '+',
-            bottom_center: '+',
-            top_right: '+',
-            middle_right: '+',
-            bottom_right: '+',
-            top_horizontal: '-',
-            middle_horizontal: '-',
-            bottom_horizontal: '-',
-            left_vertical: '|',
-            center_vertical: '|',
-            right_vertical: '|',
-
-            separate_header: true,
-            separate_rows: true,
-
-            print_left_border: true,
-            print_right_border: true,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn thin() -> Theme {
-        Theme {
-            top_left: '┌',
-            middle_left: '├',
-            bottom_left: '└',
-            top_center: '┬',
-            center: '┼',
-            bottom_center: '┴',
-            top_right: '┐',
-            middle_right: '┤',
-            bottom_right: '┘',
-
-            top_horizontal: '─',
-            middle_horizontal: '─',
-            bottom_horizontal: '─',
-
-            left_vertical: '│',
-            center_vertical: '│',
-            right_vertical: '│',
-
-            separate_header: true,
-            separate_rows: true,
-
-            print_left_border: true,
-            print_right_border: true,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn light() -> Theme {
-        Theme {
-            top_left: ' ',
-            middle_left: '─',
-            bottom_left: ' ',
-            top_center: ' ',
-            center: '─',
-            bottom_center: ' ',
-            top_right: ' ',
-            middle_right: '─',
-            bottom_right: ' ',
-
-            top_horizontal: ' ',
-            middle_horizontal: '─',
-            bottom_horizontal: ' ',
-
-            left_vertical: ' ',
-            center_vertical: ' ',
-            right_vertical: ' ',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: true,
-            print_right_border: true,
-            print_top_border: false,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn compact() -> Theme {
-        Theme {
-            top_left: '─',
-            middle_left: '─',
-            bottom_left: '─',
-            top_center: '┬',
-            center: '┼',
-            bottom_center: '┴',
-            top_right: '─',
-            middle_right: '─',
-            bottom_right: '─',
-            top_horizontal: '─',
-            middle_horizontal: '─',
-            bottom_horizontal: '─',
-
-            left_vertical: ' ',
-            center_vertical: '│',
-            right_vertical: ' ',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: false,
-            print_right_border: false,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn with_love() -> Theme {
-        Theme {
-            top_left: '❤',
-            middle_left: '❤',
-            bottom_left: '❤',
-            top_center: '❤',
-            center: '❤',
-            bottom_center: '❤',
-            top_right: '❤',
-            middle_right: '❤',
-            bottom_right: '❤',
-            top_horizontal: '❤',
-            middle_horizontal: '❤',
-            bottom_horizontal: '❤',
-
-            left_vertical: ' ',
-            center_vertical: '❤',
-            right_vertical: ' ',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: false,
-            print_right_border: false,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn compact_double() -> Theme {
-        Theme {
-            top_left: '═',
-            middle_left: '═',
-            bottom_left: '═',
-            top_center: '╦',
-            center: '╬',
-            bottom_center: '╩',
-            top_right: '═',
-            middle_right: '═',
-            bottom_right: '═',
-            top_horizontal: '═',
-            middle_horizontal: '═',
-            bottom_horizontal: '═',
-
-            left_vertical: ' ',
-            center_vertical: '║',
-            right_vertical: ' ',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: false,
-            print_right_border: false,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn rounded() -> Theme {
-        Theme {
-            top_left: '╭',
-            middle_left: '├',
-            bottom_left: '╰',
-            top_center: '┬',
-            center: '┼',
-            bottom_center: '┴',
-            top_right: '╮',
-            middle_right: '┤',
-            bottom_right: '╯',
-            top_horizontal: '─',
-            middle_horizontal: '─',
-            bottom_horizontal: '─',
-
-            left_vertical: '│',
-            center_vertical: '│',
-            right_vertical: '│',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: true,
-            print_right_border: true,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn reinforced() -> Theme {
-        Theme {
-            top_left: '┏',
-            middle_left: '├',
-            bottom_left: '┗',
-            top_center: '┬',
-            center: '┼',
-            bottom_center: '┴',
-            top_right: '┓',
-            middle_right: '┤',
-            bottom_right: '┛',
-            top_horizontal: '─',
-            middle_horizontal: '─',
-            bottom_horizontal: '─',
-
-            left_vertical: '│',
-            center_vertical: '│',
-            right_vertical: '│',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: true,
-            print_right_border: true,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn heavy() -> Theme {
-        Theme {
-            top_left: '┏',
-            middle_left: '┣',
-            bottom_left: '┗',
-            top_center: '┳',
-            center: '╋',
-            bottom_center: '┻',
-            top_right: '┓',
-            middle_right: '┫',
-            bottom_right: '┛',
-            top_horizontal: '━',
-            middle_horizontal: '━',
-            bottom_horizontal: '━',
-
-            left_vertical: '┃',
-            center_vertical: '┃',
-            right_vertical: '┃',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: true,
-            print_right_border: true,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-    #[allow(unused)]
-    pub fn none() -> Theme {
-        Theme {
-            top_left: ' ',
-            middle_left: ' ',
-            bottom_left: ' ',
-            top_center: ' ',
-            center: ' ',
-            bottom_center: ' ',
-            top_right: ' ',
-            middle_right: ' ',
-            bottom_right: ' ',
-
-            top_horizontal: ' ',
-            middle_horizontal: ' ',
-            bottom_horizontal: ' ',
-
-            left_vertical: ' ',
-            center_vertical: ' ',
-            right_vertical: ' ',
-
-            separate_header: false,
-            separate_rows: false,
-
-            print_left_border: false,
-            print_right_border: false,
-            print_top_border: false,
-            print_bottom_border: false,
-        }
-    }
+    pub theme: TableTheme,
 }
 
 impl Table {
-    pub fn new(headers: Vec<StyledString>, data: Vec<Vec<StyledString>>, theme: Theme) -> Table {
+    pub fn new(
+        headers: Vec<StyledString>,
+        data: Vec<Vec<StyledString>>,
+        theme: TableTheme,
+    ) -> Table {
         Table {
             headers,
             data,
@@ -615,7 +37,7 @@ impl Table {
 pub struct ProcessedTable {
     pub headers: Vec<ProcessedCell>,
     pub data: Vec<Vec<ProcessedCell>>,
-    pub theme: Theme,
+    pub theme: TableTheme,
 }
 
 #[derive(Debug)]
@@ -629,7 +51,7 @@ pub struct WrappedTable {
     pub column_widths: Vec<usize>,
     pub headers: Vec<WrappedCell>,
     pub data: Vec<Vec<WrappedCell>>,
-    pub theme: Theme,
+    pub theme: TableTheme,
     pub footer: Vec<WrappedCell>,
 }
